test(popover): migrate arrow
